Show current Tshirt's properties when editing
    The OptionListTagHelper class was edited in order to be able to show
the current properties of the Tshirt model being modified/edited. Before
that commit all the <option> elements inside the <select> element were
shown in the order they were rendered.
    Because that modification, when editing a Tshirt entity, at the
moment a <option> element's value is equal to one of the values of the
entity being edited, the <option> element receives a "selected"
attribute so that the current Tshirt's properties are shown. In order to
know whether the value is equal or not the custom "option-model-id"
select's attribute of OptionListTagHelper class is used. The values
being compared are GUID keys.

Mudanças a serem submetidas:
	modified:   Infrastructure/TagHelpers/OptionListTagHelper.cs
	modified:   Views/AdminTshirt/CreateAndModifyTshirtPartialView.cshtml

(use "git reset HEAD <file>..." to unstage)
lib/test.rb
~
~
".git/COMMIT_EDITMSG" 14L, 297C