Make adjusts in order to improve the code
The process between the View and the Controller of creating and
editing a Tshirt and the process of updating a Tshirt at the DAL
were improved in a way that access less the DAL and reduced the
code.

The changes were at the following places:
- AdminTshirtController.ConvertToTshirtViewModel method: the code
here was improved.
- EFTshirtRepository.SaveTshirt method: the code here was improved and
reduced to just one line at the update fase.
- To All Models at the Models Folder: all Models's properties received
a default keyword in order to be initialized automatically with a value.

No ramo optimization
Mudanças a serem submetidas:
	modified:   Controllers/AdminTshirtController.cs
	modified:   DAL/EFTshirtRepository.cs
	modified:   Models/Clothing.cs
	modified:   Models/Color.cs
	modified:   Models/Size.cs
	modified:   Models/Tshirt.cs
	modified:   Models/Type.cs
	modified:   Views/AdminTshirt/CreateAndModifyTshirtPartialView.cshtml
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.Views.dll
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.Views.pdb
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.deps.json
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.dll
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.pdb
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.runtimeconfig.dev.json
	deleted:    bin/Debug/netcoreapp2.1/Camisetas.runtimeconfig.json